refactor: add category snack bar and use it in category package
